+++ ffigen/realm_dart_collections.h
@@ -1,1 +1,0 @@
-../src/realm_dart_collections.h