+++ lib/src/realm_class.dart
@@ -37,6 +37,7 @@ export 'list.dart' show RealmList, RealmListOfObject, RealmListChanges;
 export 'realm_object.dart' show RealmEntity, RealmException, RealmObject, RealmObjectChanges;
 export 'realm_property.dart';
 export 'results.dart' show RealmResults, RealmResultsChanges;
+export 'credentials.dart' show Credentials, AuthProvider;
 
 /// A [Realm] instance represents a `Realm` database.
 ///
@@ -52,13 +53,10 @@ class Realm {
 
   /// Opens a `Realm` using a [Configuration] object.
   Realm(Configuration config) : _config = config {
-    if (_config.isInUse) {
-      throw RealmStateError("A Realm instance for this configuraiton object already exists.");
-    }
-    _scheduler = Scheduler(_config, close);
+    _scheduler = Scheduler(close);
 
     try {
-      _handle = realmCore.openRealm(_config);
+      _handle = realmCore.openRealm(_config, _scheduler);
 
       for (var realmClass in _config.schema) {
         final classMeta = realmCore.getClassMetadata(this, realmClass.name, realmClass.type);
@@ -70,7 +68,6 @@ class Realm {
       _scheduler.stop();
       rethrow;
     }
-    _config.isInUse = true;    
   }
 
   /// Deletes all files associated with a `Realm` located at given [path]
@@ -164,22 +161,22 @@ class Realm {
     }
   }
 
-  bool get _isInTransaction => realmCore.getIsWritable(this);
+  /// Checks whether the `Realm` is in write transaction.
+  bool get isInTransaction => realmCore.getIsWritable(this);
 
   /// Synchronously calls the provided callback inside a write transaction.
   ///
   /// If no exception is thrown from within the callback, the transaction will be committed.
   /// It is more efficient to update several properties or even create multiple objects in a single write transaction.
   T write<T>(T Function() writeCallback) {
+    final transaction = Transaction._(this);
+
     try {
-      realmCore.beginWrite(this);
       T result = writeCallback();
-      realmCore.commitWrite(this);
+      transaction.commit();
       return result;
     } catch (e) {
-      if (_isInTransaction) {
-        realmCore.rollbackWrite(this);
-      }
+      transaction.rollback();
       rethrow;
     }
   }
@@ -191,7 +188,6 @@ class Realm {
   void close() {
     realmCore.closeRealm(this);
     _scheduler.stop();
-    _config.isInUse = false;
   }
 
   /// Checks whether the `Realm` is closed.
@@ -258,7 +254,7 @@ class Scheduler {
   final void Function() onFinalize;
   final RawReceivePort receivePort = RawReceivePort();
 
-  Scheduler(Configuration config, this.onFinalize) {
+  Scheduler(this.onFinalize) {
     receivePort.handler = (dynamic message) {
       if (message == SCHEDULER_FINALIZE_OR_PROCESS_EXIT) {
         onFinalize();
@@ -271,12 +267,38 @@ class Scheduler {
 
     final sendPort = receivePort.sendPort;
     handle = realmCore.createScheduler(Isolate.current.hashCode, sendPort.nativePort);
-
-    realmCore.setScheduler(config, handle);
   }
 
   void stop() {
     receivePort.close();
+  }
+}
+
+/// @nodoc
+class Transaction {
+  Realm? _realm;
+
+  Transaction._(Realm realm) {
+    _realm = realm;
+    realmCore.beginWrite(realm);
+  }
+
+  void commit() {
+    if (_realm == null) {
+      throw RealmException('Transaction was already closed. Cannot commit');
+    }
+
+    realmCore.commitWrite(_realm!);
+    _realm = null;
+  }
+
+  void rollback() {
+    if (_realm == null) {
+      throw RealmException('Transaction was already closed. Cannot rollback');
+    }
+
+    realmCore.rollbackWrite(_realm!);
+    _realm = null;
   }
 }